(svn r10208) -Codechange: replace int32 with Money where appropriate.

+++ src/aircraft_cmd.cpp
@@ -493,7 +493,7 @@ CommandCost CmdSellAircraft(TileIndex tile, uint32 flags, uint32 p1, uint32 p2)
 		DoDeleteAircraft(v);
 	}
 
-	return CommandCost(-(int32)v->value);
+	return CommandCost(-v->value);
 }
 
 /** Start/Stop an aircraft.
